backend/requirements.txt
Flask>=3.0.0
Flask-Cors>=4.0.0
python-dotenv>=1.0.0
pymongo[srv]>=4.6.0
APScheduler>=3.10.4
PyJWT>=2.9.0

